[mutable] Fixed incorrect total Mutable texture size reporting due to an int32 overflow. #jira UE-205298 #rnx
#tests PIE



#virtualized

[CL 31043241 by alexei lebedev in ue5-main branch]

+++ Engine/Plugins/Experimental/Mutable/Source/CustomizableObject/Private/MuCO/CustomizableObjectSystem.cpp
@@ -3348,7 +3348,7 @@ int32 UCustomizableObjectSystem::GetTotalInstances() const
 	return NumInstances;
 }
 
-int32 UCustomizableObjectSystem::GetTextureMemoryUsed() const
+int64 UCustomizableObjectSystem::GetTextureMemoryUsed() const
 {
 	return GetPrivateChecked()->LogBenchmarkUtil.TextureGPUSize.GetValue();
 }
